Add ChallangeListPage

+++ src/services/firestoreCrudService.ts
@@ -1,19 +1,10 @@
-import RestResponse, {successResponse} from "../models/RestResponse";
+import {errorResponse, successResponse} from "../models/RestResponse";
 import Identifiable from "../models/Identifyable";
 import {addDoc, collection, deleteDoc, doc, getDoc, getDocs, updateDoc} from "firebase/firestore";
 import {db} from "./firebase";
+import CrudService from "./CrudService";
 
-
-export interface CrudService<T> {
-    path: string;
-    create: (e: T) => RestResponse<T>;
-    read: (id: string) => RestResponse<T>;
-    readAll: () => RestResponse<T[]>;
-    update: (e: T) => RestResponse<T>;
-    delete: (e: T | string) => RestResponse<unknown>;
-}
-
-export const crudService = <T extends Identifiable>(col: string): CrudService<T> => {
+export const firestoreCrudService = <T extends Identifiable>(col: string): CrudService<T> => {
     return {
         path: col,
         create: async (e) => {
@@ -42,14 +33,9 @@ export const crudService = <T extends Identifiable>(col: string): CrudService<T>
         },
         delete: async (e) => {
             const id = typeof e === "string" ? e : e.id;
+            if (id === undefined) return errorResponse(400)
             await deleteDoc(doc(db, col, id));
             return successResponse(e);
         },
     }
-}
-
-export const database = {
-    users: (id: string) => ({
-        habits: crudService(`users/${id}/habits`),
-    })
 }
